:truck: (interface): Rename from DacTimer to BasicTimer
